requirements.txt
Flask>=2.0.0
Flask-Cors>=3.0.10
requests>=2.25.1
python-dotenv>=0.15.0
httpx>=0.23.0
gunicorn>=20.1.0
click>=8.0.0
